Consolidate SalesOrderRejection and SalesOrderSuccess into SalesOrderStatus
- create SalesOrderStatus and spec file
- move tests from SalesOrderRejection and SalesOrderSuccess specs to new spec file
- update all references to either outdated class to SalesOrderStatus
- delete classes and spec files

[#127766739]
